requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
boto3==1.36.2
botocore==1.36.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
fastapi==0.115.6
frozenlist==1.5.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
idna==3.10
jmespath==1.0.1
multidict==6.1.0
oauthlib==3.2.2
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==1.26.20
websockets==14.1
yarl==1.18.3
